Review #33: decode InflightAdjustment float values
When the function byte has the high bit set, 4 bytes encode a float
adjustment value. Was reading and discarding them, hardcoding value
to 0. Now decoded as f32 and stored.

+++ propwash-core/src/format/bf/types.rs
@@ -181,7 +181,7 @@ pub enum BfEvent {
     },
     InflightAdjustment {
         function: u8,
-        value: i32,
+        value: f64,
     },
     LoggingResume {
         log_iteration: u32,
